// src/app/shared/models/export/ExportSummaryData.ts
export interface ExportSummaryData {
    totalExportOpen: string;
    totalOrderOpen: string;
    totalBtbOpen: string;
    totalExportAmount: string;
    totalOrderAmount: string;
    totalBtbAmount: string;
    totalBtbOsAmount: string;
    totalBtbBillAmount: string;
    totalBtbPayAmount: string;
    totalBtbPadAmount: string;
    totalBtbPadOsAmount: string;
    totalExOrPCAmount: string;
    totalExOrBillAmount: string;
    totalExOrPLAmount: string;
    totalExOrDisbAmount: string;
    totalExEdfRecAmount: string;

}